Nothing achieved
Trying to work out how to get a new user to be created with an Auth Token in Insomnia. Instructions are not easy to find and the ones that I found and tried are not clear and do not refer to using Insomnia.

Unticked all of the file changes, so hopefully be back to where I started